finish category operations
